Change: getBooks Either.onRight type to BookData
Previously, the onRight type was Book. This caused an error when the
retrofit client would try to map the /books response to Book. Books is
just an item in the JSON object and exists as an array under the key
"results". This has been changed to BookData, therefore, solving the
issue