Update: home screen syncing feature
This feature was to update the value of 'isBookmarked' field for each
book being displayed in BooksHomeScreen, when an delete operation is
performed in the FavouritesScreen that displays bookmarks. Previously,
the LaunchedEffect in BooksHomeScreen was using the backstackEntry as a
key which caused it to relaunch every time there was a navigation to a
different screen. This was inefficient since the BookViewModel
syncBookmarkedBooksInBookHomeScreen would be invoked every time.
Additionally, it caused delays in UI. This has been chaged to using
rememberUpdatedState with currentScreen as the LaunchedEffect key. In
this way, it will be relaunched only when the currentScreen changes.